requirements.txt
botocore~=1.31.13
py2neo~=2021.2.3
boto3~=1.28.13
neo4j~=5.10.0
requests~=2.31.0
PyYAML~=6.0.1